Finish project
Unfortunately, I did not create any other commits throughout the
creation of this project. It is a small project however, so not much was
needed. The idea was to create several html pages, and a node server to
serve these pages based on the appropriate url. I also added some
JavaScript to direct the user more easily to the other html pages by
clicking on the various icons and buttons.